backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
dataclasses-json==0.6.3
distro==1.9.0
dnspython==2.4.2
docx2txt==0.8
ecdsa==0.18.0
email-validator==2.1.0.post1
emoji==2.9.0
exceptiongroup==1.2.0
fastapi==0.108.0
filetype==1.2.0
fire==0.5.0
frozenlist==1.4.1
greenlet==3.0.1
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
httpx-sse==0.4.0
idna==3.6
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-experimental==0.0.49
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langchainhub==0.1.14
langdetect==1.0.9
langserve==0.0.46
langsmith==0.1.23
lxml==5.0.0
markupsafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.2
openai==1.10.0
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
paramiko==3.4.0
pgvector==0.2.4
pillow==10.1.0
playwright==1.40.0
pluggy==1.4.0
# psycopg2==2.9.9
psycopg2-binary
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pyee==11.0.1
pyjwt==2.8.0
pynacl==1.5.0
pysocks==1.7.1
pytest==8.0.2
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.1.2
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2024.1
pyyaml==6.0.1
rank-bm25==0.2.2
rapidfuzz==3.6.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlalchemy==2.0.27
sse-starlette==1.8.2
starlette==0.32.0.post1
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.5.2
tomli==2.0.1
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
types-requests==2.31.0.20240218
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2024.1
ujson==5.9.0
unstructured==0.11.8
unstructured-client==0.15.1
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zhipuai==2.0.1
GitPython
